database_init/requirements.txt
psycopg2-binary==2.9.9
boto3==1.35.36
urllib3==2.2.3
